requirements.txt
streamlit
pandas
python-dotenv
google-generativeai
pdfplumber
docx2txt
PyMuPDF
tiktoken
numpy
mineru[core]